requirements.txt
python>=3.8
python-dotenv>=1.0.0
pytest>=7.0.0
matplotlib>=3.5.0